refactor: convert edge:drivers:devices command to yargs
