+++ packages/edge/src/__tests__/lib/commands/drivers-util.test.ts
@@ -225,112 +225,3 @@ test('chooseInstalledDriver presents user with list of drivers with names', asyn
 	expect(apiHubdevicesListInstalledMock).toHaveBeenCalledTimes(1)
 	expect(apiHubdevicesListInstalledMock).toHaveBeenCalledWith('hub-id')
 })
-
-describe('getDriverDevices', () => {
-	it('includes only edge device types', async () => {
-		const hubDevices = [
-			{ deviceId: 'hub-device-id', label: 'Hub Label' },
-		] as Device[]
-		const edgeDevices = [
-			{
-				type: DeviceIntegrationType.LAN,
-				deviceId: 'lan-device-id',
-				label: 'LAN Device',
-				lan: {
-					driverId: 'lan-driver-id',
-					hubId: 'hub-device-id',
-				},
-			},
-			{
-				type: DeviceIntegrationType.MATTER,
-				deviceId: 'matter-device-id',
-				label: 'Matter Device',
-				matter: {
-					driverId: 'matter-driver-id',
-					hubId: 'hub-device-id',
-				},
-			},
-			{
-				type: DeviceIntegrationType.ZIGBEE,
-				deviceId: 'zigbee-device-id',
-				label: 'Zigbee Device',
-				zigbee: {
-					driverId: 'zigbee-driver-id',
-					hubId: 'hub-device-id',
-				},
-			},
-			{
-				type: DeviceIntegrationType.ZWAVE,
-				deviceId: 'zwave-device-id',
-				label: 'Z-Wave Device',
-				zwave: {
-					driverId: 'zwave-driver-id',
-					hubId: 'bad-hub-device-id',
-				},
-			},
-		] as Device[]
-		const driverDevices: driversUtil.DeviceDriverInfo[] = [
-			{
-				type: DeviceIntegrationType.LAN,
-				label: 'LAN Device',
-				deviceId: 'lan-device-id',
-				driverId: 'lan-driver-id',
-				hubId: 'hub-device-id',
-				hubLabel: 'Hub Label',
-			},
-			{
-				type: DeviceIntegrationType.MATTER,
-				label: 'Matter Device',
-				deviceId: 'matter-device-id',
-				driverId: 'matter-driver-id',
-				hubId: 'hub-device-id',
-				hubLabel: 'Hub Label',
-			},
-			{
-				type: DeviceIntegrationType.ZIGBEE,
-				label: 'Zigbee Device',
-				deviceId: 'zigbee-device-id',
-				driverId: 'zigbee-driver-id',
-				hubId: 'hub-device-id',
-				hubLabel: 'Hub Label',
-			},
-			{
-				type: DeviceIntegrationType.ZWAVE,
-				label: 'Z-Wave Device',
-				deviceId: 'zwave-device-id',
-				driverId: 'zwave-driver-id',
-				hubId: 'bad-hub-device-id',
-				hubLabel: undefined,
-			},
-		]
-
-		apiDevicesListMock.mockResolvedValueOnce(hubDevices)
-		apiDevicesListMock.mockResolvedValueOnce(edgeDevices)
-
-		expect(await getDriverDevices(client)).toStrictEqual(driverDevices)
-
-		expect(apiDevicesListMock).toHaveBeenCalledTimes(2)
-		expect(apiDevicesListMock).toHaveBeenCalledWith({ type: DeviceIntegrationType.HUB })
-		expect(apiDevicesListMock).toHaveBeenCalledWith({ type: driversUtil.edgeDeviceTypes })
-	})
-
-	it('throws exception for invalid device input', async () => {
-		const hubDevices = [{ deviceId: 'hub-device-id', label: 'Hub Label' }] as Device[]
-		const edgeDevices = [
-			{
-				type: DeviceIntegrationType.LAN,
-				deviceId: 'lan-device-id',
-				label: 'LAN Device',
-			},
-		] as Device[]
-
-		apiDevicesListMock.mockResolvedValueOnce(hubDevices)
-		apiDevicesListMock.mockResolvedValueOnce(edgeDevices)
-
-		await expect(getDriverDevices(client)).rejects.toThrow('unexpected device type LAN or missing type info')
-
-		expect(apiDevicesListMock).toHaveBeenCalledTimes(2)
-		expect(apiDevicesListMock).toHaveBeenCalledWith({ type: DeviceIntegrationType.HUB })
-		expect(apiDevicesListMock).toHaveBeenCalledWith({ type: driversUtil.edgeDeviceTypes })
-	})
-})
